backend/requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.3.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-extensions==4.1
django-filter==23.3
django-jazzmin==3.0.1
django-jet-reboot==1.3.10
django-semantic-admin==0.6.6
django-semantic-forms==0.1.8
django-unfold==0.65.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
gunicorn==23.0.0
idna==3.11
kombu==5.5.4
mysql-connector-python==9.4.0
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.10.1
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-magic
rarfile==4.2
redis==5.0.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
setuptools
